services/text-extraction/requirements.txt
# Text Extraction Service Dependencies

# PDF extraction
PyMuPDF==1.23.0

# DOCX extraction
python-docx==1.1.0

# Image OCR
pytesseract==0.3.10
Pillow==10.1.0

# Google Cloud
google-cloud-storage==2.10.0

# Supabase
supabase==2.0.0

# Utilities
python-dateutil==2.8.2
